+++ src/useFieldArray.ts
@@ -2,6 +2,7 @@ import * as React from 'react';
 
 import focusFieldBy from './logic/focusFieldBy';
 import getFocusFieldName from './logic/getFocusFieldName';
+import isWatched from './logic/isWatched';
 import mapCurrentIds from './logic/mapCurrentIds';
 import mapIds from './logic/mapId';
 import appendAt from './utils/append';
@@ -236,7 +237,10 @@ export const useFieldArray = <
       index,
       value,
     );
-    _fieldIds.current = mapIds(updatedFieldArrayValues, keyName);
+    _fieldIds.current = mapIds<TFieldValues, TFieldArrayName, TKeyName>(
+      updatedFieldArrayValues,
+      keyName,
+    );
     control._updateFieldArray(
       name,
       updateAt,
@@ -256,9 +260,11 @@ export const useFieldArray = <
       | Partial<FieldArray<TFieldValues, TFieldArrayName>>
       | Partial<FieldArray<TFieldValues, TFieldArrayName>>[],
   ) => {
-    const updatedFieldArrayValuesWithKey: Partial<
-      FieldArrayWithId<TFieldValues, TFieldArrayName, TKeyName>
-    >[] = mapIds(convertToArrayPayload(value), keyName);
+    const updatedFieldArrayValuesWithKey = mapIds<
+      TFieldValues,
+      TFieldArrayName,
+      TKeyName
+    >(convertToArrayPayload(value) as Partial<TFieldValues>[], keyName);
     control._updateFieldArray(
       name,
       () => updatedFieldArrayValuesWithKey,
@@ -273,16 +279,7 @@ export const useFieldArray = <
   React.useEffect(() => {
     control._stateFlags.action = false;
 
-    if (control._names.watchAll) {
-      control._subjects.state.next({});
-    } else {
-      for (const watchField of control._names.watch) {
-        if (name.startsWith(watchField)) {
-          control._subjects.state.next({});
-          break;
-        }
-      }
-    }
+    isWatched(name, control._names) && control._subjects.state.next({});
 
     if (_actioned.current) {
       control._executeSchema([name]).then((result) => {