+++ src/useForm.ts
@@ -48,11 +48,11 @@ export function useForm<
 
   useSubscribe({
     subject: control._subjects.state,
-    callback: (formState) => {
-      if (shouldRenderFormState(formState, control._proxyFormState, true)) {
+    callback: (value) => {
+      if (shouldRenderFormState(value, control._proxyFormState, true)) {
         control._formState = {
           ...control._formState,
-          ...formState,
+          ...value,
         };
 
         updateFormState({ ...control._formState });