+++ src/useFormState.ts
@@ -24,6 +24,7 @@ function useFormState<TFieldValues extends FieldValues = FieldValues>(
       : (CONTROL_DEFAULT as unknown as Control<TFieldValues>),
     disabled,
     name,
+    exact
   } = props || {};
   const [formState, updateFormState] = React.useState(control._formState);
   const _localProxyFormState = React.useRef({
@@ -40,15 +41,16 @@ function useFormState<TFieldValues extends FieldValues = FieldValues>(
 
   useSubscribe({
     disabled,
-    callback: (formState) =>
+    callback: (value) =>
       shouldSubscribeByName(
         _name.current as InternalFieldName,
-        formState.name,
+        value.name,
+        exact,
       ) &&
-      shouldRenderFormState(formState, _localProxyFormState.current) &&
+      shouldRenderFormState(value, _localProxyFormState.current) &&
       updateFormState({
         ...control._formState,
-        ...formState,
+        ...value,
       }),
     subject: control._subjects.state,
   });